+++ packages/moonstone/Scrollable/Scrollable.js
@@ -12,6 +12,7 @@ import {constants, ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollabl
 import {getDirection} from '@enact/spotlight';
 import {getTargetByDirectionFromElement, getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import {Job} from '@enact/core/util';
+import platform from '@enact/core/platform';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import ReactDOM from 'react-dom';
@@ -226,6 +227,10 @@ class ScrollableBase extends Component {
 	lastScrollPositionOnFocus = null
 	indexToFocus = null
 	nodeToFocus = null
+
+	// voice control
+	isVoiceControl = false
+	voiceControlDirection = 'vertical'
 
 	// overscroll
 	overscrollRefs = {horizontal: null, vertical: null}
@@ -501,6 +506,10 @@ class ScrollableBase extends Component {
 		this.lastFocusedItem = null;
 		this.lastScrollPositionOnFocus = null;
 		this.isWheeling = false;
+		if (this.isVoiceControl) {
+			this.isVoiceControl = false;
+			this.updateFocusAfterVoiceControl();
+		}
 	}
 
 	focusOnItem () {
@@ -596,6 +605,10 @@ class ScrollableBase extends Component {
 	addEventListeners = (childContainerRef) => {
 		if (childContainerRef && childContainerRef.addEventListener) {
 			childContainerRef.addEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.addEventListener('webOSVoice', this.onVoice);
+				childContainerRef.setAttribute('data-webos-voice-intent', 'Scroll');
+			}
 		}
 	}
 
@@ -603,6 +616,10 @@ class ScrollableBase extends Component {
 	removeEventListeners = (childContainerRef) => {
 		if (childContainerRef && childContainerRef.removeEventListener) {
 			childContainerRef.removeEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.removeEventListener('webOSVoice', this.onVoice);
+				childContainerRef.removeAttribute('data-webos-voice-intent');
+			}
 		}
 	}
 
@@ -632,6 +649,71 @@ class ScrollableBase extends Component {
 		if (ref) {
 			this.uiRef = ref;
 		}
+	}
+
+	updateFocusAfterVoiceControl = () => {
+		const spotItem = Spotlight.getCurrent();
+		if (spotItem && this.uiRef.containerRef.contains(spotItem)) {
+			const
+				viewportBounds = this.uiRef.containerRef.getBoundingClientRect(),
+				spotItemBounds = spotItem.getBoundingClientRect(),
+				nodes = Spotlight.getSpottableDescendants(this.uiRef.containerRef.dataset.spotlightId),
+				first = this.voiceControlDirection === 'vertical' ? 'top' : 'left',
+				last = this.voiceControlDirection === 'vertical' ? 'bottom' : 'right';
+
+			if (spotItemBounds[last] < viewportBounds[first] || spotItemBounds[first] > viewportBounds[last]) {
+				for (let i = 0; i < nodes.length; i++) {
+					const nodeBounds = nodes[i].getBoundingClientRect();
+					if (nodeBounds[first] > viewportBounds[first] && nodeBounds[last] < viewportBounds[last]) {
+						Spotlight.focus(nodes[i]);
+						break;
+					}
+				}
+			}
+		}
+	}
+
+	onVoice = (e) => {
+		const scroll = e && e.detail && e.detail.scroll;
+		this.isVoiceControl = true;
+
+		switch (scroll) {
+			case 'up':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: true});
+				break;
+			case 'down':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: true});
+				break;
+			case 'left':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: false});
+				break;
+			case 'right':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: false});
+				break;
+			case 'top':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'top'});
+				break;
+			case 'bottom':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'bottom'});
+				break;
+			case 'leftmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'left'});
+				break;
+			case 'rightmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'right'});
+				break;
+			default:
+				this.isVoiceControl = false;
+		}
+		e.preventDefault();
 	}
 
 	render () {