+++ packages/sampler/stories/qa-stories/VirtualList.js
@@ -28,7 +28,7 @@ const
 	renderItem = (size) => ({index, ...rest}) => {
 		const itemStyle = {height: size + 'px', ...style.item};
 		return (
-			<StatefulSwitchItem  index={index} style={itemStyle} {...rest}>
+			<StatefulSwitchItem index={index} style={itemStyle} {...rest}>
 				{items[index].item}
 			</StatefulSwitchItem>
 		);
@@ -58,7 +58,9 @@ class StatefulSwitchItem extends React.Component {
 
 	onToggle = () => {
 		items[this.props.index].selected = !items[this.props.index].selected;
-		this.setState({selected: !this.state.selected});
+		this.setState(({selected}) => ({
+			selected: !selected
+		}));
 	}
 
 	render () {